+++ src/kit/Changelog.md
@@ -3,6 +3,7 @@
 ## Version 0.7.5
 * Add a documentation is available :D Yeah !
 * Rename UiEvent to EventUi
+* Add a hook in App before mount views
 * EventUi is now a static function (not parent anymore by Ui components)
 * Some language code in UtilsI18n have been change regarding of what Sony said
 * Split CmakeLists.txt to be more readable and easy to configure